services/bot/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
redis==5.0.1
python-dotenv==1.0.0
openai==1.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
